+++ docs-demo/demos/VirtualList/index.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import mockjs from 'mockjs';
-import { StkTableColumn } from '../../../../src/StkTable/index';
-import StkTable from '../../../StkTable.vue';
+import { StkTableColumn } from '@/StkTable/index';
+import StkTable from '../../StkTable.vue';
 import Panel from './Panel.vue';
 import type { DataType } from './types';
 
@@ -17,7 +17,7 @@ const data = [
     },
     ...new Array(20).fill(0).map((_, i) => ({
         id: i,
-        title: mockjs.Random.csentence(5, 15),
+        title: mockjs.Random.ctitle(5, 15),
         content: mockjs.Random.cparagraph(1, 10),
         date: mockjs.Random.datetime('yyyy-MM-dd HH:mm'),
     })),