+++ docs-src/.vitepress/config.ts
@@ -131,6 +131,7 @@ export default defineConfig({
         base: '/demos',
         items: [
           { text: '大量数据', link: '/huge-data' },
+          { text: '虚拟单列表', link: '/virtual-list' },
         ]
       }
     },