+++ lib/src/StkTable/useColResize.d.ts
@@ -1,5 +1,6 @@
 import { ComputedRef, Ref, ShallowRef } from 'vue';
 import { StkTableColumn, UniqKey } from './types';
+
 type Params<DT extends Record<string, any>> = {
     props: any;
     emits: any;