+++ lib/src/StkTable/useGetFixedColPosition.d.ts
@@ -1,5 +1,6 @@
 import { ComputedRef, ShallowRef } from 'vue';
 import { StkTableColumn, UniqKey } from './types';
+
 type Params<T extends Record<string, any>> = {
     colKeyGen: ComputedRef<(col: StkTableColumn<T>) => UniqKey>;
     tableHeadersForCalc: ShallowRef<StkTableColumn<T>[][]>;