+++ lib/src/StkTable/useThDrag.d.ts
@@ -1,5 +1,6 @@
 import { ComputedRef } from 'vue';
 import { StkTableColumn, UniqKey } from './types';
+
 type Params<T extends Record<string, any>> = {
     props: any;
     emits: any;