+++ lib/src/StkTable/useTrDrag.d.ts
@@ -1,5 +1,6 @@
 import { ShallowRef } from 'vue';
 import { DragRowConfig } from './types';
+
 type Params = {
     props: any;
     emits: any;