+++ lib/src/StkTable/useVirtualScroll.d.ts
@@ -1,5 +1,6 @@
 import { Ref, ShallowRef } from 'vue';
 import { StkTableColumn, UniqKey } from './types';
+
 type Option<DT extends Record<string, any>> = {
     props: any;
     tableContainerRef: Ref<HTMLElement | undefined>;